monitor/dbmw: 审计 A 类批量修复——ReplyUserInfo 成员键按服务 #ifdef(+0xc guild/+0x32 monitor)、CharLogin 日志原文、MemberSecede m_type 反转、GameServerRegist 键、MemberEnterReply/RequestMemberEnter 日志参数、PvPChannelUserCount 写包、Tcp线程异常文本、ItemLimitEdition 有符号比较、SaveGuildWarPointList ||

+++ dnf_decompile/source/DNFServer/GameServer/Monitor/ItemLimitEditionManager.cpp
@@ -152,7 +152,7 @@ void CItemLimitEditionMgr::processScheduledJob(CApplication* app, bool flag)
         for (std::map<unsigned int, CItemLimitEdition*>::const_iterator it = m_items.begin();
              it != m_items.end(); ++it)
         {
-            if (it->second->getSellEndTime() < (unsigned int)now)
+            if ((int)it->second->getSellEndTime() < (int)now)
             {
                 expired[count] = it->second->getIPGNO();
                 count++;
